+++ veterans/keystone-fallback.d.ts
@@ -18,7 +18,6 @@ declare module '.keystone/types' {
     export namespace Source {
       export type TypeInfo<Session = any> = any;
     }
-
     export namespace Region {
       export type TypeInfo<Session = any> = any;
     }