+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - regionsList component ([117](https://github.com/boarlabsxyz/razom/pull/117))
 - Google Auth ([118](https://github.com/boarlabsxyz/razom/pull/118))
 - reset filter button to the map section ([128](https://github.com/boarlabsxyz/razom/pull/128))
+- region filed to create initiative page, get the list from db to display on web-site ([141](https://github.com/boarlabsxyz/razom/pull/141))
 
 ### Changed